miden_lifted_stark::lmcs::bitrev

Trait BitReversibleMatrix

pub trait BitReversibleMatrix<T: Send + Sync + Clone>: Matrix<T> {
    type BitRev: BitReversibleMatrix<T>;

    // Required method
    fn bit_reverse_rows(self) -> Self::BitRev;
}

A matrix that supports bit-reversed row reordering.

Local copy of p3_matrix::bitrev::BitReversibleMatrix extended with impls for [FlatMatrixView].

Required Associated Types

type BitRev: BitReversibleMatrix<T>

The type returned when this matrix is viewed in bit-reversed order.

Required Methods

fn bit_reverse_rows(self) -> Self::BitRev

Return a version of the matrix with its row order reversed by bit index.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.